give ErrorBar its own demo file
